requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
